requirements.txt
Django==3.2.9
djangorestframework==3.12.4
django-cors-headers==3.8.0
django-filter==2.4.0
django-querycount==0.7.0
django-rest-passwordreset==1.2.1
Pillow==8.4.0
psycopg2==2.9.2
django-rest-auth==0.9.5
djangorestframework-jwt==1.11.0
requests==2.27.1
gunicorn==20.1.0
whitenoise==6.0.0
django-allauth
django-tenant-schemas==1.11.0
django-allauth==0.49.0